Add files via upload
This is the capstone project finished my MIT - Applied Data Science Bootcamp - Apr 2021.

Objectives:
Analyze and build a machine learning model to help identify which clients are more likely to default on their loan.
Find the factors driving the loan defaulted chance.
Create a profile of the clients who are likely to default. 
Give recommendations to the bank on the important features to consider while approving  a loan.

The key questions:
What are the important features of clients to consider while approving a loan?
Which model should be considered the best model to predict the loan default?
What should the bank do enhance the best practices for risk management?

The problem formulation:
Identify important features driving to loan default probabilities.
Build a good model to predict a client who are likely to be a loan default.
Using these predictions, our model then help the banks to allocate effective their resources, minimize the risk and produce a good internal rate of return (IRR).